pyproject.toml
[tool.poetry]
name = "django-graphql"
version = "0.1.0"
description = ""
authors = ["Ivan Zugnoni <zugnoni.ivan@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
django = "^3.2.4"
django-model-utils = "^4.1.1"
graphene-django = "^2.15.0"
django-filter = "^2.4.0"

[tool.poetry.dev-dependencies]
ipdb = "^0.13.9"
black = "^19.10b0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
